+++ modeloParcial/src/modeloParcial/Coco.java
@@ -2,15 +2,15 @@ package modeloParcial;
 
 public class Coco extends Bacteria{
 	
-	private int radio;
+	private double radio;
 
-	public Coco(int radio, int edad) {
+	public Coco(double radio, int edad) {
 		super(edad);
 		this.radio = radio;
 	}
 
 	@Override
-	public Integer consumoO2() {
+	public double consumoO2() {
 		return 2 * this.radio;
 	}
