+++ modeloParcial/src/modeloParcial/Consumidor.java
@@ -2,7 +2,7 @@ package modeloParcial;
 
 public interface Consumidor {
 	
-	public Integer consumoO2();
+	public double consumoO2();
 
 	public boolean esInmediata();
