+++ modeloParcial/src/modeloParcial/Poblacion.java
@@ -5,21 +5,21 @@ import java.util.List;
 
 public class Poblacion implements Consumidor{
 	
-	private int factor;
+	private double factor;
 	private List<Consumidor> consumidores;
 	
-	public Poblacion(int factor) {
+	public Poblacion(double factor) {
 		this.factor = factor;
 		this.consumidores = new ArrayList<Consumidor>();
 	}
 
 	@Override
-	public Integer consumoO2() {
+	public double consumoO2() {
 		return (this.sumaConsumoComponentes() * this.cantidadBacteriasInmediatas()) / this.factor;
 	}
 
-	private int cantidadBacteriasInmediatas() {
-		int cantidad = 0;
+	private double cantidadBacteriasInmediatas() {
+		double cantidad = 0;
 		for (int i = 0; i < this.consumidores.size(); i++) {
 			Consumidor consumidor = this.consumidores.get(i);
 			cantidad = cantidad + unoSiEsInmediata(consumidor);
@@ -27,15 +27,15 @@ public class Poblacion implements Consumidor{
 		return cantidad;
 	}
 
-	private int unoSiEsInmediata(Consumidor consumidor) {
+	private double unoSiEsInmediata(Consumidor consumidor) {
 		if (consumidor.esInmediata()) {
 			return 1;
 		}
 		return 0;
 	}
 
-	private int sumaConsumoComponentes() {
-		int sumaTotal = 0;
+	private double sumaConsumoComponentes() {
+		double sumaTotal = 0;
 		for (int i = 0; i < this.consumidores.size(); i++) {
 			Consumidor consumidor = this.consumidores.get(i);
 			sumaTotal = sumaTotal + consumidor.consumoO2();
